Implement correct opacity and filter handling within backdrop-filter
Prior to this CL, opacity and filters applied to the backdrop-filtered
element did not get applied to the backdrop-filtered backdrop. Now, the
opacity and filters get applied directly to the backdrop-filtered
image directly, which is an approximation but should be very close.

Also fixed up a few of the backdrop-filter tests to account for this,
and removed a few lines from TestExpectations that now work.

Bug: 497522, 524689
Change-Id: If671bdf65ac0ae04cf175143eb8949ff4be0837b